dashboard done

+++ src/Dashboard/RestuarantsDash/RestPages/Logout/Dashboardheader.jsx
@@ -1,21 +1,42 @@
 import { Nav, NavDropdown } from "react-bootstrap";
 import "../../../../Pages/pages.css";
 import Logo from "../../../../assets/Logo.png";
-import { useNavigate } from "react-router-dom";
+import { NavLink, useNavigate } from "react-router-dom";
 
-const dashboardHeader = () => {
+const DashboardHeader = () => {
   const navigate = useNavigate();
   let user = JSON.parse(localStorage.getItem("User"));
   function logout() {
-    localStorage.clear();
-    navigate("/login");
+    Swal.fire({
+      title: "Are you sure?",
+      text: "",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, Logout",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          title: "Logged Out",
+          text: "You have been successfully logged out.",
+          icon: "success",
+          timer: 2000, // Set a timer for 5000 milliseconds (5 seconds)
+          showConfirmButton: false, // Hide the confirm button
+        });
+        localStorage.clear();
+        navigate("/login");
+      }
+    });
   }
   return (
     <div>
-      <header className="head" style={{ height: "6vh" }}>
+      <header className="head">
         <section className="headwrap">
-          <div className="Logodiv">
-            <img src={Logo} alt="" />
+          <div className="Logodiv" style={{ cursor: "pointer" }}>
+            <NavLink to="/">
+              <img src={Logo} alt="" />
+            </NavLink>
           </div>
           <div className="signindiv">
             <Nav style={{ padding: "10px" }}>
@@ -44,4 +65,4 @@ const dashboardHeader = () => {
   );
 };
 
-export default dashboardHeader;
+export default DashboardHeader;
